requirements.txt
Bootstrap-Flask==2.1.0
click==8.1.3
dominate==2.7.0
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
PyYAML==6.0
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
flask-sqlalchemy
flask-wtf